+++ kit/lib/routing.js
@@ -3,12 +3,21 @@
 // ----------------------
 // IMPORTS
 
+/* NPM */
 import React from 'react';
 import PropTypes from 'prop-types';
 
+// Browser history, that we can use to control URL pushstate throughout our
+// entire app
+import createBrowserHistory from 'history/createBrowserHistory';
+
+// React Router
 import { Route, Redirect as ReactRouterRedirect } from 'react-router-dom';
 
 // ----------------------
+
+// Create and export a custom history
+export const history = !SERVER && createBrowserHistory();
 
 // <Status code="xxx"> component.  Updates the context router's context, which
 // can be used by the server handler to respond to the status on the server.